+++ tests/run-almost-all.ts
@@ -137,6 +137,7 @@ import { testit as issue6045 } from "./issue-6045.ts";
 import { testit as issue6046 } from "./issue-6046.ts";
 import { testit as issue6048 } from "./issue-6048.ts";
 import { testit as issue6053 } from "./issue-6053.ts";
+import { testit as issue6054 } from "./issue-6054.ts";
 import { testit as favoritesTabAfterDocumentClose } from "./favorites-tab-after-document-close.ts";
 import { testit as movePolygonPolylineInk } from "./move-polygon-polyline-ink.ts";
 import { testit as annotDeleteRedraw } from "./annot-delete-redraw.ts";
@@ -239,6 +240,7 @@ export const tests: NamedTest[] = [
   ["issue-6046", issue6046],
   ["issue-6048", issue6048],
   ["issue-6053", issue6053],
+  ["issue-6054", issue6054],
   ["move-polygon-polyline-ink", movePolygonPolylineInk],
   ["annot-delete-redraw", annotDeleteRedraw],
   ["annot-icon-over-contents", annotIconOverContents],